Actually refactor windows buttons out to separate widget
